feat: move logic to the NetworkList feature
