+++ src/renderer/features/networks/NetworkList/index.ts
@@ -1,0 +1,1 @@
+export { networkListModel } from './model/network-list';
