+++ src/renderer/features/networks/index.ts
@@ -1,1 +1,2 @@
 export { filterModel } from './NetworkFilter';
+export { networkListModel } from './NetworkList';
